SupperMarket-Backend/requirements.txt
fastapi==0.95.1
pymongo==4.3.3
python-dotenv==1.0.0
